Refactor how proxy settings are passed to Spark
This allows us to not repeat the WMF specific hostnames
and rely instead on configuration passed in.

If using wmfdata in WMF Jupyter, the proxy settings will
be set and propagated and used locally and on remote workers.

If using wmfdata outside of WMF Jupyter, you have to make sure the
local http proxy environment variables are set.