spark.py - keep check_kerberos_auth for all sessions
This is what was done before my recent changes, and on second thought
is the better behavior.  Even if not running Spark in YARN, it is
likely that users will be using Spark to access data in HDFS, so
we might as well just always enforce kinit before using spark.