Ensure binary data in MariaDB is returned as strings
Previously, we relied on the fact that mysql-connector-python
would convert any binary data to strings whenever the connection
was set to use the UTF-8 character set. This behavior was changed
in version 8.0.24, so that binary data is returned as bytearrays
regardless of the specified character set (although the setting
still causes field names to be converted).

This change adds a converter class that explicitly converts any
binary data to strings.