Update the ca_bundle used for presto connections
Until now we have been using Puppet certificates to encrypt connections
to Presto. We will shortly be using certificates that are generated by
the PKI system instead.

As such, wmfdata-python needs to know about our root CA as well as the
puppet CA. This change enables that by using a new, combined CA bundle.

Once this change in in place, we can safely update the presto
certificates and wmfdata-python will continue to work.

Bug: T345482