Allow the use of a DNS CNAME for analytics-presto
We wish to use a CNAME for analytics-presto.eqiad.wmnet in order to make
it possible to migrate the presto coordinator role from one host to
another.

Kerberos, however, makes this a bit tricky because the coordinator still
uses its FQDN for the hostname part of the principal that is uses to
authenticate itself to clients.

In order to work around this issue, this commit adds a DNS lookup step
to the presto class, so that it will be able to know the hostname of the
coordinator when it attempts authentication. This matches the behaviour
of the presto client.

Bug: T345482