Remove the code to disable hostname warnings from urllib3
We have switched presto so that it uses PKI certificates instead of
puppet certificates now. As a result of this, the certificates generated
also have SubjectAlternativeName attributes which contain the fully
qualified DNS name of the host.

Therefore, we can remove this custom code that disables the warnings
that the attribute is not present with a matching value.

Bug: T345482